+++ packages/types/src/types/unlockTypes.ts
@@ -41,6 +41,7 @@ export interface Token {
 export enum HookType {
   CUSTOM_CONTRACT = 'CUSTOM_CONTRACT',
   PASSWORD = 'PASSWORD',
+  PROMOCODE = 'PROMOCODE',
 }
 
 export const HooksName = [